Function rust_dgc::get_payload

pub fn get_payload(raw_payload: &[u8]) -> Result<CwtParsed, Box<dyn Error>>

Get the CWT payload, which according to RFC-8392 Section-7 is just the Cose_Sign1 Struct from RFC-8152 Section-4.2 This means we have a CBOR-Array like this, where the Headers defines the Algorithm used and payload contains CWT’s claims. In our case the hcert will be one of those claims.

COSE Sign struct

COSE_Sign = [
      Headers,
      payload : bstr / nil,
      signatures : [+ COSE_Signature]
  ]

COSE Protected Headers and Unprotected Headers

Headers = (
      protected : empty_or_serialized_map,
      unprotected : header_map
  )
  header_map = {
      Generic_Headers,
      * label => values
  }

  empty_or_serialized_map = bstr .cbor header_map / bstr .size 0